infoschema: Migrate `dataForMetricTables` from package `infoschema` to `executor`
